move out pure map state suites
